#958: refactoring - create package util.quarter.generate
