+++ backend/src/test/java/ch/puzzle/okr/service/persistence/QuarterPersistenceServiceIT.java
@@ -5,7 +5,7 @@ import ch.puzzle.okr.exception.OkrResponseStatusException;
 import ch.puzzle.okr.models.Quarter;
 import ch.puzzle.okr.test.SpringIntegrationTest;
 import ch.puzzle.okr.test.TestHelper;
-import ch.puzzle.okr.util.QuarterRangeChecker;
+import ch.puzzle.okr.util.quarter.check.QuarterRangeChecker;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
